USART : - corrected syntax error in if statement. - changed the parameter int type from unsigned 16bit to 32bit.

+++ lib/usart.h
@@ -13,7 +13,7 @@
  * PC2 -> RX
  * But can be changed using syscfg 
  */
-void usart_init(uint16_t BAUDRATE);
+void usart_init(uint32_t BAUDRATE);
 
 void usart_write(uint8_t data);
